services/data-generator/requirements.txt
kafka-python==2.0.2
pymongo==4.6.0
pandas==2.1.3
numpy==1.24.3
faker==20.1.0